feat: unlimited implementation of meter attribute getter

+++ src/Cryptlex.LexFloatClient/LicenseMeterAttribute.cs
@@ -7,15 +7,27 @@ namespace Cryptlex
     public class HostLicenseMeterAttribute
     {
 
+        /// <summary>
+        /// The name of the meter attribute.
+        /// </summary>
         public string Name;
 
-        public uint AllowedUses;
+        /// <summary>
+        /// The allowed uses of the meter attribute. A value of -1 indicates unlimited allowed uses.
+        /// </summary>
+        public long AllowedUses;
 
-        public uint TotalUses;
+        /// <summary>
+        /// The total uses of the meter attribute.
+        /// </summary>
+        public ulong TotalUses;
 
-        public uint GrossUses;
+        /// <summary>
+        /// The gross uses of the meter attribute.
+        /// </summary>
+        public ulong GrossUses;
 
-        public HostLicenseMeterAttribute(string name, uint allowedUses, uint totalUses, uint grossUses)
+        public HostLicenseMeterAttribute(string name, long allowedUses, ulong totalUses, ulong grossUses)
         {
             this.Name = name;
             this.AllowedUses = allowedUses;
